Update Linux build

+++ tests/test_launcher.cpp
@@ -51,9 +51,9 @@ protected:
         std::this_thread::sleep_for(std::chrono::milliseconds(context.z()));
         phase3->StopPhase();
 
-        context.metrics().SetCustom("AAA", 123ll);
-        context.metrics().SetCustom("BBB", 456ll);
-        context.metrics().SetCustom("CCC", 789ll);
+        context.metrics().SetCustom("AAA", (int64_t)123);
+        context.metrics().SetCustom("BBB", (int64_t)456);
+        context.metrics().SetCustom("CCC", (int64_t)789);
         context.metrics().SetCustom("AAA", 1.23);
         context.metrics().SetCustom("BBB", 4.56);
         context.metrics().SetCustom("CCC", 7.89);
